Added overlay.prototype.fixWidth to improve box sizing after scroll
Just as with our original YUI-based panel, we have a problem when overlays
are rendered past initial viewport because text is wrapped.

CSS `width: max-content` would be the ideal fix, but isn't supported by
IE/Edge (yet).

Instead, we move the overlay to position `0, 0`, determine the width
without word-wrapping, fix the width, and then move the overlay back
to its original position.